feat: split out scroll light client types from unionlabs
