+++ lib/unionlabs/src/ibc/lightclients.rs
@@ -2,5 +2,4 @@ pub mod ethereum;
 pub mod evm_in_cosmos;
 pub mod linea;
 pub mod movement;
-pub mod scroll;
 pub mod wasm;
